requirements.txt
django
wheel
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
django-rest-framework
django-cors-headers